Changed the existing product version mapping from working on the resource itself into a recursive call that will update its children as well. This is to overcome the issue in that a platform was getting its reference created correctly but none of its direct child servers that were discovered.
Also added a call to create the product version in the updateResourceVersion method. Admittedly, I'm not sure why this is a separate method instead of done in the resource data merge, but it sounds like a place where, since the resource version is changing, the product version needs to be updated as well.