JBNADM-3198 - The following changes are somewhat temporary for JON 2.0. I see the ultimate solution moving towards this approach, but this is currently limiting from a strictly generic point of view: - For creating a package-backed resource, default the version to timestamp to prevent the user from having to manage the version strings themselves. The version input isn't even displayed to the user. For the EAR/WAR case, this makes more sense than showing it to them. - For deploying a new package on a package-backed resource, we currently lock the user into only being able to create packages of the backing package type. We also use the existing package's name/architecture/type. In essence, for a package-backed resource, we transform the deploy new package flow into more of an update-like workflow.
Again, this is limiting if plugins were to use the package functionality more extensively, but for JON 2.0 (and specifically the EAR/WAR case), this is a good medium for now. This is also closer to what I'd like to see package-backed resource updates looking like in the future. We can split this out into a separate part of the UI post JON-2.0.