Adding support for login() and logout() functions to used in scripts.
Initial commit for Controller.java which defines the login and logout
methods that are exposed to scripts. Controller is bound to scripts
as 'rhq'. The commands have been refactored to take a Controller
instance. LoginCommand and LogoutOut command now delegate to
Controller. Overloaded versions of login() will likely come in a
future commit. For now, you can do

    rhq.login('rhqadmin', 'rhqadmin')

which logs you in against localhost on the default port of 7080.
The defaults can be changed as follows,

    rhq.server.host = 'myhost';
    rhq.server.port = myport;
    rhq.login('rhqadmin', 'rhqadmin');