Fixing mapping typos in RawConfiguration and adding JPA associatin in Configuration.
Adding a couple tests in ConfigurationTest.java to ensure that the association
is mapped correctly. Also updating tests to make sure that the JTA transaction
is started before I get the entity manager. Getting the entity manager first
and then starting the transaction led to the following exception in one case,

javax.persistence.TransactionRequiredException: no transaction is in progress