Second interim checking for raw config editing
Made RawConfigCollection conversation scoped, driven off of annotations.
Added an 'init' method to start the conversation, because seam doesn't allow
starting a conversation based on a constructor.
There is still a bug with one file showing up as modified after a commit
The commit code does not yet send code to the back end.
In order to show behavior, the mock code is enabled.  This will be removed before going live, probably when the back save gets enabled.