Fixing failing tests, disabling one, debugging another.
The failure in ConfigurationTest was a legitimate failure due to a bug in
RawConfiguration.deepCopy(). The failure in PluginTest was a result of trying
to use bad data. Adding some debugging to ScriptAssertTest as I have a hunch
that the javax.script libraries are not installed with the openjdk packages
that are being used on the build server.