[US 403] Initial commit for PluginTransformer, PluginTransformerTest, and PluginTransformerException
PluginTransformer takes a PluginDescriptor and a URL for the plugin JAR file and converts them
into a Plugin object. This logic is needed by PluginManager when the plugin container is used in
embedded mode in order to build the list of loaded plugins. That list was previously a list of
just plugin names, but now it is a list of the Plugin domain objects. In enterprise mode, we get
those plugins from the server, but in embedded mode, we have to create the Plugin objects
ourselves since there is no server to talk to.