[US 403] Modifying plugin container to load Plugin objects in their entirety.
Adding pluginsOnServer property to PluginContainerConfiguration. This set contains plugins sent
down from the server. It is only populated when the plugin container is running inside an agent.

Modyfing PluginUpdate to populate PluginContainerConfiguration.pluginsOnServer with the plugins
sent down from the server.

Updating PluginManager to store loaded plugin as Plugin objects. Previously, only the names
were stored. Now we can more easily retrive the ampsVersion of a plugin. Support has been added
to load plugins both in enterripse (i.e., running inside an agent) and in embedded mode.

Updating ProductPluginDeployer to set the ampsVersion of the Plugin object that is creates.