[US 2] Adding configurationFormat attribute to plugin descriptor schema.
Adding some build support for Groovy code. This is the initial commit for PluginMetadataParserTest
which is written in Groovy. I decided to write the test in Groovy for improved readability. Each
of the test methods involves exercising the parser class for some plugin descriptor. Groovy's
enhanced string capabilities allows me to embed the plugin descriptor XML write into the test as
opposed to referencing some external file.