[US 403] Refactoring PluginManager to better support integration tests.
Initial commit for PluginManagerTest which has tests to verify that plugins are successfully
loaded, to verify the order in which they are loaded, and to verify the order in which they are
shut down.

Refactoring the code in PluginManager that loads and manages the cache of PluginLifecycleListener
objects. Thay code is now encapsulated behind the PluginLifecycleListenerManager interface. This
allowed me to introduce a test hook to verify certain conditions.