Adding additional configuration to agent and comm tests to allow multiple builds run simultaneously on hudson.
The agent tests use a couple config files. I have provided alternate files for each of those, and filtering is
done on those alternate files. Distinct filters are defined in the settings-rawconfig.xml and in the
settings-content.xml in order to avoid port conflicts. AgentCommTestBase has been modified so that when either of
those two settings file is used, the alternate agent config files will be used.