[US 2] Initial commit to support raw config in ConfigurationManager
ConfigurationManager will continue supporting the "legacy" ConfigurationFacet as well as the new
ResourceConfigurationFacet. The strategy pattern is being utilized to better encapsulate branching logic as
well as to improve testability. The branching logic to determine the strategy is handled by a
ResourceConfigurationStrategyFactory (its implementing class, ResourceConfigurationStrategyFactoryImpl).

Adding support for unit testing with JMock. Initial commit for JMockTest which is a base test class that
is intended to behave similar to the JMock JUnit4 runner in that it handles intializing a Mockery object
for use in tests and it also invokes Mockery.assertIsSatisfied after a test run.