Commenting out failing test for now as the functionality may be removed.
For now PluginManager has been modified to always read and load the Plugin objects from disk instead of
fetching them from the server. I am not sure if we actually need to get them from the server. If we do not
then these tests need to be updated accordingly to reflect the changes.