[US 2] Updating agent serialization to include the rawConfigurations collection in the (de)serialization.
Adding unit tests for serialization/deserialization for when the externalization strategy is AGENT. Also
changing how the properties map is copied. It is now creating a copy by calling the deepCopy() method of each
Property and adding that property to a new map. The same is done for the RawConfigurations.

Changing another place in ConfigurationManagerBean to call Configuration.deepCopyWithoutProxies() instead of
calling Configuration.deepCopy(false).