[US 6] Refactoring common test classes into test-utils module.
Moving JMockTest into test-utils so that it can be resued across different modules.

Initial commit for AssertUtils which for now contains only a single method,
assertPropertyMatches(). This method compare all accessible properties of two
objects. Internally, that method is implemented by org.rhq.test.PropertyMathcher.

Initial commit for org.rhq.test.jmock.PropertyMatcher. This is a custom JMock
matcher that performs matching based on the accessible properties of an object.