[US 6] Starting to refactor UpdateResourceConfigurationRunner for raw config support and lots of refactoring.
Renaming LoadResourceConfiguration to ConfigManagement since the interface now
provides operations both for loading and updating configs. Renaming
implementors to more logical names.

Refactoring UpdatResourceConfigurationRunner to take an instance of
ConfigManagement instead of ConfigurationFacet.

Adding the method assertCollectionEqualsNoOrder() in AssertUtils in the test-utils
module. The method implementation is largely delegated to CollectionEqualsChecker
has logic to provide a useful error report. It will generate a message that tells
you if elements were found in the collection that should not have been there as
well as if elements were not found that should have been.