Fixing build failures in webservices module
We were getting compiler errors in some of the test classes in the webservices module after
I added the new configuration update operation to ConfigurationManagerRemote. It caused a
problem because the method is overloaded and SOAP does not support method overloading. I
resolved the issue by overriding the default value of the generated SOAP operation name
to be updateStructuredOrRawResourceConfiguration().