1156 - Fix excessive login's to RHN when downloading packages/files We were calling a login per file/pkg download, this resulted in the systemid being flagged as abusive from RHN.  Fix is to call login once and save the auth credentials for subsequent calls.
In addition this commit touched some of the RHNProvider unit tests.  We added
a means to do an exhaustive check of package syncs but downloading each package and
comparing it's md5sum to the expected md5sum.  This is a time intensive test and is
controlled by a java property defined in the pom.xml, by default this test is off.