[US 1025] Fixing raw config plugin redeployment error
When redeploying a plugin with raw config, I was getting a NPE in ConfigurationDefinition
while ConfigurationMetadataManagerBean is applying the update. The NPE happens because of
a couple of the fields in ConfigurationDefinition which are Maps are lazily initialized.
Initializing them at the time of instantiation (of the ConfigurationDefinition object)
to empty maps resolves the issue.