[BZ 538088] Reverting back to using the view ui bean which only has a request scope.
view.xhtml in several places in the JSF code was making calls to
ExistingResourceConfigurationUIBean which has session scope. As a result,
external config updagtes were not immediately visible because the configuration
objects were getting cached for the duration of the session, whereas
ExistingResourceConfigurationViewUIBean has a scope of request so every request
for the view page results in fetching the latest configuration for the resource.