[BZ 537891] Fixing bug in which required query parameter was not getted appended to url
The resource id is expected to be a parameter in the request. When on edit-raw.xhtml and
clicking the link to go back to structured, the resource id basically gets dropped and
then navigating to any other page, tab, etc. that expects the id parameter results in
an exception. I added a new navigation rule similar to one already defined for
edit.xhtml (structured edit page) that adds the id to the request and does a server
side redirect.