Refactoring facelets code so that the appropriate ui is rendered for the config type
Refactoring view.xhtml to support both structured and raw. With these changes, we now only
see structured UI for a config that supports only structured, we only see the raw ui for
a config that supports only raw, and we see both structured and raw for a config that
supports that both.