[BZ 537891] Adding fix for edge case in which id parameter was not passed when leaving raw edit page.
When editing a structured config (that also supports raw) and navigating to the raw edit page
via the switch to raw link, the resource id parameter was not getting passed properly. As a
result, when navigating to another tab for that resource from the raw edit page would result
in an exception since the id parameter was null. Adding a navigation rule so that the parameter
is passed.