Seam-ifying backing bean in order to work around request scope issue with JSF
On view.xhtml, when the user clicks the links to change back and forth between structured
and raw, it results in a postback with an action method invoked to perform the translation.
A problem I encountered was that once the view rendered after the postback the backing bean
which was in request scope was no longer available to the view. Consequetly, clicking any
commandLink or commandButton did not invoke any of the specified actions. Making the backing
bean a Seam component with page scope resolved this issue.