Adding logic to group files by directory in the file menu
Right now this logic is only in place on the view page. Initial commit for RawConfigDirectory which is just
a container for raw config files to group them by directory. This is the initial commit for
RawConfigUIBean as well which provides properties used in the view, like displaying the short name
of the config file for instance.