Fixing regression that broke translation and changing files from the raw file menu
I introduced a regression earlier today when adding support for the modal file editor that resulted in
breaking the translation when switching back and forth between the structured and raw tabs. Selecting
a file in the raw file menu triggers a partial page update. The modal editor is in a subview on the page and
it was not getting updating during the ajax request. It was holding onto an older value of the selected
file which in turn was overwriting the corresponding property in the backing bean.