Adding logic to properly initialize the rawConfigDirectories property in the UI backing bean.
The rawConfigDirectories property is only used for configs that support raw so a check needed to be
added to only initialize it if raw is supported. Previously we were attempting to always
initialize it, and when raw was not supported it caused a NPE. When raw is not supported it
gets initialized to an empty collection since it is annotated with @Out and Seam will not
allow an @Out to be null.