Fixing a bug that breaks translating from structured to raw
There has been a lingering bug in the UI bean code that appears to break translating from
structured to raw. An initial translation from structured to raw works fine. It happens on
subsequent translations.

The UI bean mantains a cache of modified raw config values. When translating from structured
to raw, the cache of modified values was not getting updated correctly. It was retaining
stale values. Then on the raw edit (Advanced) tab, when you select a file for editing that
has a property that was modified on structured side, the UI bean code first looks in the cache
of modified values. If the raw config file has a modified value in the cache, that value is
rendered. The problem though was that the cache had stale values and should have been getting
flushed when translating from structured to raw the translated values are now newer than
whats in the cache.