Adding save/cancel buttons to modal editor
Clicking the cancel button simply closes the editor as a client-side operation. Clicking
the save button closes the editor but also fires an ajax request to the server to update
the on-page text editor. The modification will result in the file in the file menu
appearing as being updated as well.