Modifying upload functionality so that the upload component is now rendered in a modal window
The upload dialog/modal window is rendered on the edit page when you click the 'Upload New' link.
Once you upload the file, the modal window closes, the contents of the file in text area are
replaced with the contents of the uploaded file, and the file will be marked as being modified
in the file menu.