Updating logic for key events to indicate whether or not a file has been modified
When you start editing the contents of a raw config file, an event get fired off to the
server to determine if the edits have modified the file. If so, the file menu is updated
to reflect that the file has been modified. If you further edit the file to undo you
edits, the file menu will get updated to reflect that the file has been restored to
its original state.