Updating logic for displaying the number of files changed
The number of files changed was getting lost between translations since the modified cache
get purged on a translation call. I added a counter to keep track of the number of changed
files.

I also had to update some of the richfaces code on edit.xhtml so that both the upper and
the lower labels get rerendered when the file is edited.