Initial refactoring to replace usage of ExistingResourceConfigurationViewUIBean with configurationViewer.
ExistingResourceConfigurationViewUIBean was written as a JSF backing bean. configurationViewer
is a seam component instance of the class ConfigurationEditor. It is a port of JSF backing that
leverages seam capabilities like bijection and pre-render actions.

As for the bijection, the configuration and configuration definition are now out-jected into
the context under the names of resourceConfiguration and resourceConfigurationDefinition,
making them available to view.xhtml

Deleting xhtml files that are no longer used. Changing the logging level of a logging statement
in ResourceTreeModelUIBean to debug to reduce noise in the logs.

Changing the page flow behavior when viewing a structured configuration with complex properties
and then clicking on the view button to view a list of maps property. Previously, the request
was forwarded on the server side and the view rendered is view-map.xhtml. Since
configurationViewer has page scope, we lose state when view-map.xhtml is rendered. Now a
client-side redirect to view-map.xhtml is performed with configurationViewer.initialize()
being invoked in a pre-render action which allows for the configuration and configuration
definition to be properly initialized.