Starting seam refactoring for edit pages
Two main changes with this commit. First, updating the logic for the view.xhtml page action such that
the pre-render action is not invoked on post backs. Secondly, when you click the edit button, you
are taken to edit.xhtml and remain on the same tab whereas previously the navigation logic there
was hard coded. Still need to update the logic for the lower edit button.