Refactoring methods for translating between structured and raw and for tracking changes
As part of this refactoring, introducing a new class, ResourceConfigurationViewer, which
ResourceConfigurationEditor extends. ResourceConfigurationEditor has methods to handle
translating between structured and raw.

RawConfigUIBean has been refactored such that it encapsulates logic to determinr whether
or not the corresponding raw config file is modified. When the file is modified,
RawConfigUIBean fires a seam event to inidicate that the file has been updated.

ResourceConfigurationEditor subscribes to the event fired by RawConfigUIBean. The event
handler method contains logic for tracking the files that are modified.