Change "LIKE :filter"  logic in NamedQuery to "LIKE :filter ESCAPE :escapeChar" logic.
This brings oracle and postgres (and any others) into consistent behavior
while allowing flexibility with the escape char setting.  Note that we
now consistently treat SQL wildcard chars, % and _, as literals in the
specified :filter value. In addition, occurrences of the escape character
will also be treated as a literal.