Adding and updating support for undo operation in ResourceConfigurationEditor and in edit.xhtml
The undo operation through the file menu now works for resources that support structured and
raw configuration. The code changes still need to be applied to raw-only. I had to refactor
the file menu JSF code a bit. First, the file menu item is now omitted and has been replaced
with a <h:outputLabel>. The undo button was previously rendered with <h:commandButton> and
used <a4j:support>. It is now rendered with <aj4:button>.

These changes allow for each component to respond separately to mouse clicks and the a4j
button allows for an ajax request to be sent instead of a full post-back. Ajax requests
should be possible using,

  <h:commandButton>
    <a4j:support/>
  </h:commandButton>

but for some reason it was causing full post-backs. I also encounted some seam-related
exceptions involving concurrent request timeouts for the same conversation. It is not
clear why those exceptions were getting thrown, but they were causing the undo operatin
to break. Specifying the concurrent-request-timeout attribute for the edit.xhtml page
navigation declaration in pages.xml resolved that issue.