Refactoring common code in edit.xhtml into shared templates
Renaming view-structured.xhtml to structured.xhtml as it is now used in both view.xhtml and in
edit.xhtml. Initial commit for raw.xhtml which is used in both view.xhtml and in edit.xhtml.
Updating rerendering logic for the modal editor so that making changes to the file in the
modal editor updates the file menu. And when clicking the undo button in the file menu, the
modal editor contents are also reset.