Fixing file upload
There were a handful of changes needed to get things working. First, the upload component
needed to be in a form of its own. Secondly, the form has to be nested inside the modal
panel; otherwise, the upload listener is not called. I am not sure why that happens, but
this was reported in https://jira.jboss.org/jira/browse/RF-3492. Next, the resource id
cannot be submitted with the file upload request which causes problems in some places
that expect to find the resource id as a request parameter during the update model phase
of the JSF lifecycle. Two things were done to deal with this. First, the life of the
ResourceUIBean for a given request is extended using <a4j:keepAlive>. That allows us
to get ahold of the resource id from ResourceUIBean. Secondly, some code, like in
FavoritesUIBean, that was looking directly at the request for the id, was refactored to get
the id from ResourceUIBean.

The file upload code was pulled out of ResourceConfigurationEditor and put into a new seam
component, FileUploadComponent which is just a thin wrapper around FileUploadUIBean. Right
now FileUploadComponent is tightly coupled to raw config, but it can/should be refactored
so that it can be used for any file upload that implemented via richfaces and seam.