Changing file path status lable to a regular JSF label and updating some of the file menu rerendering
The status label would change to say 'Loading <path name>..." when you select a file. The
problem though is that if File 1 is currently displayed and you click on File 2, the status
lable would read, 'Loading File 1...', and then when File 2 is rendered, the label would
read 'File 2'. In order to get the status message to display, 'Loading File 2...', there
needs to be some JavaScript to update the label to the selected file. I spent a little bit
of time investigating this but did not find a solution. I decided to change it to just a
regular label that is only change once the new file is rendered. This way the label always
displays the correct name.

Updating the file menu rendering logic. On edit.xhtml, the background of the selected file
was not changing. I previously was rerendering the entire file menu when a file menu item
is selected. By not rerendering the entire file menu, the background of the selected menu
item changes. Not sure why this behaves the way it does.