Refactoring ResourceConfigurationViewer to handle a null resource config def and/or null config
When the viewer initializes it attempts to load the resource configuration definition and
resource configuration. If the definition is null, then that means that the resource does
not support configuration. If the configuration is null, it is assumed that the configuration
has not yet been initialized. This might happen for instance if the agent is down.