Adding file display name property in RawConfigUIBean so file name is not lost in download
The fileName property was modified in a previous commit to prepend an asterisk to the name
when the file has modifications. This was done to be more consistent with the mock-ups from
UXD. That change however resulted in the file name for a download request to show up as
only an asterisk because the download method was using that same property. That property
now just returns the file name, minus the path, and there is a new property, fileDisplayName
which has the logic for prepending the asterisk.