Adding logic to only start conversations when necessary and adding logic to handle session timeouts for ajax requests
When a user's session times out a ViewExpiredException is thrown. There is a framework in
place within rhq to handle by sending the user to login and then redirecting the user back
to the previous view after successful login. This functionality does not work though for
ajax requests. The redirect url gets mangled with jsf parameters and the user is not sent
to the previous view as a result.

I have utilized some richfaces support for session expiration so when an ajax request is
fired off and the session has expired, the user will get an alert that the sesison has
expired and then be redirected to the login page. There will be no after that however to
redirect the user back to the previous view.