First cut at summary counts portlet for the dashboard.
This is the initial commit for the ResourceBossGWTService and corresponding classes. Also the
initial commit for SummaryCountsView which is the widget that displays the inventory counts.
I am also adding the SummaryCounts which is just a wrapper around the class InventorySummary.
Initially I tried returning an InventorySummary object from ResourceBossService but that does
not work since it lives in the portal-war module. If it does not break anything I may wind up
moving InventorySummary into core/domain so that it is accessible client side.