Partially revert the fix for
https://bugzilla.redhat.com/show_bug.cgi?id=562236

Leave the change moving to OID from bytea for Postgres PackageBits.bits
but eliminate the use for Hibernate classes (SerializableBlob and BlobImpl).
These classes can not be used in the CLI or Agent (or anywhere for in
GWT-Aware classes).

This change goes back to using byte[] for the bits field. This seems to
work fine for the purposes of the unit tests ans scripst, so we can again
move forward with bundle deployment etc.  Whether or not this pairing
solves the issue with the BZ I'm not sure. It needs to be tested and if
its no good we need another, possible native (non-hibernate) solution
for streaming the bits back and forth.