Changing the masked property value since the unicode chars were getting stripped off
Password fields were masked with a value of "\u0002MASKED\u0003" prior to being rendered.
When the form is submitted though, the password property has the unicode characters stripped
off which in turn resulted in the password property not getting unmasked.

I stumbled onto this when updating the connection properties for a jboss server and I
realized that the value for the credentials property was getting lost.