[BZ 573034] Adding test to reproduce the bug
The registers a plugin and then creates a resource of the resource type specified in the
plugin. A new version of the plugin is deployed having a new plugin configuration property
that specifies a default value. The test then checks that the resource's plugin configuration
picks up the new property and it is set to the specified default value.