[BZ 573034] Removing unncessary call to EntityManager.remove() and EntityManager.flush()
Removing the unncessary calls to EntityManager.remove() and EntityManager.flush() when a property
definition is being removed. The call to EntityManager.remove() is unncessary because the
property definition is first removed from the configuration definition and the association
from configuration definition to property definition cascades all operations, including
deletes.

The call to EntityManager.remove() was causing an exception in
ConfigurationMetadataManagerBeanTest that reported that you cannot call EntityManager.remove()
on a detached entity. I am not clear why that this failed in the test but works in production
code; however, I tested the change in this commit against a running server and it seemed
to work fine. I redeployed a plugin with properties removed and those property definitions
were successfully removed from the database.

Lastly, I did review http://jira.rhq-project.org/browse/RHQ-575 before making the change
and have proceeded with caution in light of the comments in the jira.