[BZ 573034] Updating ConfigurationMetadataManagerBean to generate a report of property defs added
ResourceMetadataManagerBean nows gets back a report from ConfigurationMetadataManagerBean
that says which property definitions have been added to the configuration definition. All of
the resources for the given resource type are fetched and their plugin configurations are
updated with new properties using their default values for the new property definitions that
are required.