Domain Model Change and Refactor Change from a single PackageType for a BundleType to a PackageType per Bundle. The former approach originally worked with the idea that Package.classification set to bundleName could be used to distinguish between BundleFile packages with the same name but contained in different bundles.  But the fact that classification was never designed to be used as part of the natural key for a package would have forced a large refactor potentially conflicting with the current use of classification.
The alternatives were to add a Package.subType to package and make it part
of the natural key.  This would have been specific to the needs of bundling
and again added changes to the underlying content domain model.  Instead
it was deemed reasonable to have a devoted PackageType per Bundle.  This
model does create a potentially large number of packageTypes
(one per bundle) on the BundleType's ResourceType. But it's not clear that
creates any problem.  It has the advantage of leveraging the Package
design without introducing any more complexity. Also, Bundle deletion now
can cause cascade removals of the packages for its package type.