[BZ 535640] Modifying the column sizing algorithm to limit when columns will be truncated.
Previously column sizes were uniformly reduced so that all columns fit in the alloted space.
The problem with this is that it resulted in truncated columns that were not too large.
With the new algorithm all columns are initially set to the same size.  For each column that
needs space than what was allocated to it, the extra space is reallocated to other columns
that need more space. Once the extra space is calculated the column widths are recalculated
for those remaining columns that still need to be resized.