[BZ 577047] Moving project files from scripts module to new cli-tests module
Having the cli tests as a top-level module will make it easier to run
them from hudson. This way, I can avoid having to add profiles to
several poms to deactivate modules.

This commit also includes an update to the script test runner that binds
the rhq server name to the script variable, rhqServerName. The value is
obtained from the system property, rhq.server.name.