[BZ 546397] The property name for log event sources needs to match the name in the plugin descriptor
An exception was getting thrown when trying to add a new event source. The event sources
are stored as PropertyMaps in a PropertyList. PropertyList has a check to ensure that
each of the added to it have the same name. If a property doesn't then an exception is
thrown. Based on the plugin descriptor that property name should be logEventSource but
JBossASDiscoveryComponent was giving a name of serverLog. So when you go to add a new
event source, the new corresponding property is named logEventSource since that is what
is in the plugin configuration definition, but the initial property added during
discovery has is named serverLog; hence the exception.