[BZ 544353] Fixing UI bean code so stale values are not used when creating a new group def
NewGroupDefinitionGeneralPropertiesUIBean and EditGroupDefinitionGeneralPropertiesUIBean
both stick an attribute in the session to indicate that particular fields should be
cached and reused. Part of the problem was that both beans were using the same attribute
name. The other part of the problem is that during group creation, the marker attribute
was not getting removed from the session, and the absence of the marker indicates that
any cached values should not be used.