[BZ 582991] Fixing dbupgrade script so that values for context column match enum value
Based on the enum value hibernate is expecting to find the value 'RESOURCE' in
rhq_saved_search.context but instead was finding 'Resource'. Modifying upgrade script to
use all caps value to be consistent with enum. This is already done in the set up script.

Adding some default configuration for the surefire plugin. When running a test in core/domain
I was getting OutOfMemoryErrors. I am not sure what the heap settings are for the forked vm,
but it must be pretty low. Setting a default args line to pass to the forked vm that sets
the min heap size to 16 MB and the max heap size to 256 MB.