[BZ 582991] undoing deletion of operation_def_id from rhq_alert_definition since it is still referenced
Based on some disucssion with heiko we do want to delete the operation_def_id, but we also
want to remove the corresponding AlertDefinition.operationDefinition field and its usages.
operationDefinition is still used however in some obsolete code that needs to be cleaned
up. We will try to address this later all in one fell swoop.