Refactoring dbutils build to support building databases for earlier releases
Consolidated dbutils-related properties into a set of common properties and using profiles
to determine which database to use. Currently three database profiles are supported,

  1. dev - dev container
  2. test - automated tests
  3. upgrade - testing an upgrade

The antrun plugin code in dbutils/pom.xml was moved into a profile and is run by default.
When building an upgrade database you will want to use another new profile, release which
is activating with -Ddbutils.mode=release. And specifiy the release as -Drelease=2.3.1
so the ant script can determine which scripts to use.

Adding a new target to dbsetup-build.xml for creating an upgrade database.