BZ 586943 - Adding work around for hibernate exception
Hibernate exceptions are occurring because the underlying collections in PropertyMap and
PropertyList which include a cascade style of delete_orphan, are getting derefenced. I
have not been able to determine when/where the PropertyMap.map and PropertyList.list
collections are getting derefenced. This commit works around the problem by creating
a deep copy of the Configuration object, stripped of any hibernate proxies.