[BZ 586068] Moving formatting code out of core/domain to avoid problems with gwt
There are some classes and packages such as java.text that cannot be used in GWT code. One
of my previous commits refactored MeasurementOOBComposite to use MeasurementConverter which
has a dependency on java.text. I removed the formatting calls from MeausrementOOBComposite and
moved formmatting classes into server/jar. The formatting is now done in the SLSB method from
which the MeasurementOOBComposites are returned. The GWT-based formatter,
MeasurementConverterClient has been moved to coregui.