[BZ 596753] Fixing table renaming upgrade
rhq_channel, rhq_channel_resource_map, rhq_channel_content_src_map, and rhq_channel_pkg_version_map
were renamed in a commit back in October of last year. New tables were introduced but these tables
were not dropped from the schema. References to them however were removed from content-schema.xml.
So during an upgrade in which we wipe out an existing database, these tables do not get removed
which results in some key violations eventually causing the upgrade/installation to fail.

With this commit, the aforementioned channel tables have been added to obsolete_schema.xml
so that they are removed during the upgrade, thus allowing the upgrade/installation to
complete successfully.

There is a second part of this bug which is not addressed with this commit. In the event of
an upgrade in which we do not wipe out an existing db any data in the channel tables that
are being renamed is lost. The data is not copied over into the new tables. I will address
this in a subsequent commit.