BZ-599691: automatically reselect the active notification each time "SAVE" is pressed
figure out which one should be selected from the backing store.  updating
an alert notification will overwrite all existing notifications with the
latest data.  since the ids of the alert notifications are different after
a save-action, but since we know what the updated data, we can select the
correct one by testing for data equality (instead of pk/id equality).