[BZ 604291] Adding logic to handle NPEs and adding some test utility classes
A NullPointerException was being thrown when trying to print a resource's current
availability when the resource did not have its currentAvailability set. The
currentAvailability will be null if the resource has not yet been committed into
inventory and it will be null for a (presumably) short period after being committed
until its availability has been reported.

ResourceBuilder and ResourceTypeBuilder are some test utility classes that can
be used for building up object graphs in tests without obscuring the test and
simplifying the test fixture set up. These classes are a work in progress, and in
a subsequent commit they will be moved into a separate test utility module, maybe
something like domain-test-util.